# Minimize allocations in strutil