--- a/strutil/go.mod
+++ b/strutil/go.mod
@@ -2,7 +2,4 @@ module github.com/hashicorp/go-secure-stdlib/strutil
 
 go 1.16
 
-require (
-	github.com/ryanuber/go-glob v1.0.0
-	github.com/stretchr/testify v1.7.0
-)
+require github.com/ryanuber/go-glob v1.0.0